userid

+++ project 3/Models/SP_flutter_SumQuantity_Orders_Query_Result.cs
@@ -15,6 +15,6 @@ namespace project_3.Models
     {
         public string productName { get; set; }
         public string measre_name { get; set; }
-        public Nullable<int> sumQuantity { get; set; }
+        public Nullable<double> sumQuantity { get; set; }
     }
 }
